backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg[binary]==3.2.3
tzdata==2024.1
